fixed brand col migration in mobiles, laptops

+++ database/migrations/2017_02_15_190715_add_brand_id_to_mobiles_laptops.php
@@ -18,7 +18,9 @@ class AddBrandIdToMobilesLaptops extends Migration
                 $table->unsignedInteger('brand_id');
                 $table->foreign('brand_id')->references('id')->on('brands');
             }
-            $table->dropColumn('brand');
+            if(!Schema::hasColumn('mobiles', 'brand')){
+                $table->dropColumn('brand');
+            }
         });
 
         Schema::table('laptops', function(Blueprint $table){
@@ -26,7 +28,9 @@ class AddBrandIdToMobilesLaptops extends Migration
                 $table->unsignedInteger('brand_id');
                 $table->foreign('brand_id')->references('id')->on('brands');
             }
-            $table->dropColumn('brand');
+            if(!Schema::hasColumn('mobiles', 'brand')){
+                $table->dropColumn('brand');
+            }
         });
     }
